refactor: relocate extension scope + invocation-context glue to core
Continues dissolving the core<->extension coupling. Like the manifest store
(#8852), ExtensionScope and ResolvedExtensionInvocationContext are core glue,
not extension behavior: they bridge Component/Project (core) with the contract
ExtensionManifest, using the now-core manifest store to load.

Relocates scope.rs -> crate::extension_scope and invocation_context.rs ->
crate::extension_invocation_context. Repointed their only extension couplings
(super::manifest::ExtensionManifest -> homeboy_extension_contract, super::
load_extension -> crate::extension_store::load_extension, super::ExtensionScope
-> crate::extension_scope). The extension module re-exports both types so all
crate::extension::* paths stay stable with zero churn.

Both new modules are fully independent of the extension module - further
shrinking the behavior surface that blocks a severable homeboy-extension crate.
