+++ crates/homeboy-core/src/extension/mod.rs
@@ -10,7 +10,7 @@ pub mod component_script;
 mod env_provider;
 mod execution;
 mod fingerprint;
-mod invocation_context;
+// invocation_context relocated to crate::extension_invocation_context
 // The grammar parsing engine is a language-agnostic primitive; it now lives in
 // homeboy-engine-primitives. Re-exported here so existing
 // `crate::extension::grammar` / `crate::extension::grammar_items` paths keep
@@ -29,7 +29,7 @@ mod refactor_protocol;
 mod repair;
 mod runner;
 mod runtime_helper;
-mod scope;
+// scope relocated to crate::extension_scope (core glue over the contract manifest)
 pub mod self_check;
 mod setup_env;
 mod summary;
@@ -58,6 +58,8 @@ pub use homeboy_extension_contract::core_compat::{
 };
 pub use homeboy_extension_contract::ExtensionCapability;
 
+pub use crate::extension_invocation_context::ResolvedExtensionInvocationContext;
+pub use crate::extension_scope::ExtensionScope;
 pub use crate::extension_store::{
     available_extension_ids, extension_path, find_extension_by_tool, find_extension_for_file_ext,
     is_extension_linked, load_all_extensions, load_extension, merge, save_manifest,
@@ -83,7 +85,6 @@ pub use homeboy_extension_contract::update_output::{
 };
 pub use homeboy_extension_contract::version::{parse_extension_version, VersionConstraint};
 pub use homeboy_extension_contract::{DeployArchiveInstallPolicy, DeployRequiredHeader};
-pub use invocation_context::ResolvedExtensionInvocationContext;
 pub use lifecycle::source_metadata::resolve_source_url;
 pub use lifecycle::source_metadata::SourceMetadataRepair;
 pub use lifecycle::{
@@ -128,7 +129,6 @@ pub use runner::{ExtensionRunner, RunnerOutput};
 pub use runtime_helper::{
     helper_path, BASH_PREFLIGHT_ENV, COMMAND_CAPTURE_ENV, RUNNER_PRELUDE_ENV, RUNNER_STEPS_ENV,
 };
-pub use scope::ExtensionScope;
 pub use summary::{list_summaries, ActionSummary, ExtensionSummary};
 pub use validation::{
     extension_provides_build, validate_extension_requirements, validate_required_extensions,
