+++ crates/homeboy-core/src/extension_invocation_context.rs
@@ -5,7 +5,8 @@ use crate::engine::template;
 use crate::error::{Error, Result};
 use crate::project::{self, Project};
 
-use super::{ExtensionManifest, ExtensionScope};
+use crate::extension_scope::ExtensionScope;
+use homeboy_extension_contract::ExtensionManifest;
 
 /// The resolved project/component/settings identity for one extension invocation.
 ///
