+++ crates/homeboy-core/src/lib.rs
@@ -95,7 +95,9 @@ pub mod execution;
 pub mod execution_contract;
 pub(crate) mod expand;
 pub mod extension;
+pub mod extension_invocation_context;
 pub mod extension_provider_discovery;
+pub mod extension_scope;
 pub mod extension_store;
 // finding moved to the internal `homeboy-finding` crate. Re-exported so existing
 // `crate::finding::*` call sites keep working unchanged.
